changed GAtomicCounter type to gint instead of guint16 to allow window sizes larger than 65536
	This also synces syslog-ng with the 3.0 branch, where this is already gint instead of guint16.

+++ src/atomic.h
@@ -63,7 +63,7 @@ g_atomic_counter_set(GAtomicCounter *c, gint value)
 }
 
 #else
-typedef guint16 GAtomicCounter;
+typedef gint GAtomicCounter;
 
 static inline void
 g_atomic_counter_inc(GAtomicCounter *c)
